Lab9 modified : set_i's type casting changed (unsigned int -> size_t)

+++ Lab9/macros.h
@@ -12,7 +12,7 @@
 #define RANGE_DESC(curr, max, min) for ((curr) = (max); (curr) >= (min); (curr)--)
 
 #define SET(ary, start, count, value)   do { \
-                                            unsigned int set_i; \
+                                            size_t set_i; \
                                             for ((set_i) = (start); (set_i) < ((start) + (count)); (set_i)++) { \
                                                 ary[(set_i)] = (value); \
                                             } \
